requirements.txt
Django==5.1.3
djangorestframework==3.15.2
dj-database-url==2.3.0 
psycopg2==2.9.10
python-dotenv==1.0.1 
drf-spectacular==0.27.2
pytest==8.3.4
pytest-django==4.9.0